convonet/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioice==0.10.1
aiortc==1.11.0
aiosignal>=1.4.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.8.7
autogen-agentchat==0.5.3
autogen-azure==0.0.0
autogen-core==0.5.3
av==14.3.0
azure-core==1.33.0
azure-identity==1.21.0
blinker==1.8.2
Bootstrap-Flask==2.3.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dominate==2.9.1
executing==2.2.0
fastapi==0.115.12
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.6.0
google-api-python-client>=2.181.0
google-auth==2.28.2
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
google-crc32c==1.7.1
langchain-core>=0.3.0
langchain-mcp-adapters>=0.1.1
langchain-openai>=0.3.17
langgraph>=0.4.5
lxml>=5.4.0
mcp>=1.9.0
pandas>=2.2.3
psycopg2-binary>=2.9.10
greenlet==3.0.3
gunicorn==21.2.0
uvicorn[standard]>=0.20.0
asgiref>=3.6.0
Flask-SocketIO>=5.0.0
eventlet>=0.33.0
gevent>=23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
jiter==0.9.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.11.0
Mako==1.3.10
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==10.6.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
mysql-connector-python==9.3.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.76.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prance==23.6.21.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary>=2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyautogen==0.8.7
pybars4==0.9.13
pycparser==2.22
pydantic>=2.11.4
pydantic-settings==2.9.1
pydantic_core>=2.33.2
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
bcrypt==4.0.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv>=1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy>=1.15.3
semantic-kernel==1.28.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy>=2.0.41
sounddevice>=0.5.2
stack-data==0.6.3
starlette==0.46.2
termcolor==3.0.1
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
twilio>=8.0.0
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.4.0
visitor==0.1.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.17.2
WTForms==3.0.1
yarl==1.20.0
zipp==3.21.0
sentry-sdk[flask]>=2.0.0
redis>=5.0.0
# Audio processing libraries for audio stream player (optional - requires system audio libraries)
# PyAudio>=0.2.11  # Commented out - requires PortAudio system library not available on Render.com

# Additional integrations
composio-core>=0.3.0
flask-limiter>=3.0.0

# LiveKit SDK
livekit==0.17.5
livekit-api>=0.6.0

# Deepgram SDK for WebRTC speech-to-text
deepgram-sdk>=3.0.0

# Tavily web search for AI agents
tavily-python>=0.7.0
